putting get all absentees in separate file
