+++ src/firebase/firebase.js
@@ -11,6 +11,8 @@ export const auth = firebase.auth();
 
 export { getLessons, getDateLessons, writeLessons } from "./lessons";
 
+export { getAllAbsentees } from './absence';
+
 export async function writeRegistration(registration) {
   const studentDocName = getStudentDocName(registration);
   return db.collection("registrations").doc(studentDocName).set(registration)
@@ -26,7 +28,7 @@ export async function getUsersRegistrations() {
   const querySnapshot = await db.collection("registrations").where("email", "==", getUserEmail()).get();
   let registrations = [];
   querySnapshot.forEach((doc) => {
-    registrations.push(doc.data())
+    registrations.push(doc.data());
   });
   return registrations;
 }
@@ -35,7 +37,7 @@ export async function getAllRegistrations() {
   const querySnapshot = await db.collection("registrations").get();
   let registrations = [];
   querySnapshot.forEach((doc) => {
-    registrations.push(doc.data())
+    registrations.push(doc.data());
   });
   return registrations;
 }
@@ -43,14 +45,14 @@ export async function getAllRegistrations() {
 export async function createGroup(group) {
   const userIsModerator = await checkUserClaim('moderator');
   if (!userIsModerator) {
-    return new Error('User not authorized.')
+    return new Error('User not authorized.');
   }
   return db.collection("groups").doc(group.groupName).set(group)
     .then(()=> {
-      return {success: true}
+      return {success: true};
     })
     .catch((error)=> {
-      throw new Error(error)
+      throw new Error(error);
     });
 }
 
@@ -65,7 +67,7 @@ export async function getGroups() {
   const querySnapshot = await db.collection("groups").get();
   let groups = [];
   querySnapshot.forEach((doc) => {
-    groups.push(doc.data())
+    groups.push(doc.data());
   });
   return groups;
 }
@@ -79,15 +81,15 @@ export async function removeStudentFromGroups(student, groups) {
   groups.forEach((group) => {
     db.collection('groups').doc(group.groupName).collection('students').doc(studentDocName).delete()
       .catch((error) => {
-        return new Error("Error removing document: ", error)
+        return new Error("Error removing document: ", error);
       });
-  })
+  });
 }
 
 export async function writeStudentToGroup(student, groupName) {
   const userIsModerator = await checkUserClaim('moderator');
   if (!userIsModerator) {
-    return new Error('User not authorized.')
+    return new Error('User not authorized.');
   }
   const updatedStudent = Object.assign(student, {group: groupName});
   const studentDocName = getStudentDocName(student);
@@ -97,7 +99,7 @@ export async function writeStudentToGroup(student, groupName) {
 export async function updateGroupTeacher(teacherEmail, groupName) {
   const userIsModerator = await checkUserClaim('moderator');
   if (!userIsModerator) {
-    return new Error('User not authorized.')
+    return new Error('User not authorized.');
   }
   const group = await getGroup(groupName);
   const groupRef = await db.collection("groups").doc(group);
@@ -109,7 +111,7 @@ export async function updateGroupTeacher(teacherEmail, groupName) {
 export async function removeGroupTeacher(teacherEmail, groupName) {
   const userIsModerator = await checkUserClaim('moderator');
   if (!userIsModerator) {
-    return new Error('User not authorized.')
+    return new Error('User not authorized.');
   }
   const group = await getGroup(groupName);
   const groupRef = await db.collection("groups").doc(group);
@@ -121,16 +123,16 @@ export async function removeGroupTeacher(teacherEmail, groupName) {
 export async function updateRegistration(registration, groupName) {
   const userIsModerator = await checkUserClaim('moderator');
   if (!userIsModerator) {
-    return new Error('User not authorized.')
+    return new Error('User not authorized.');
   }
   const studentDocName = getStudentDocName(registration);
   const registrationRef = db.collection("registrations").doc(studentDocName);
   return registrationRef.update({
     group: groupName
   }).then(() => {
-    return {success: true}
+    return {success: true};
   }).catch((error) => {
-    return new Error('Error updating document: ', error)
+    return new Error('Error updating document: ', error);
   });
 }
 
@@ -138,10 +140,10 @@ export async function getGroupsOf(teacherEmail) {
   const querySnapshot = await db.collection("groups").where("teachers", "array-contains", teacherEmail).get();
   let teachersGroups = [];
   querySnapshot.forEach(async (doc) => {
-    const teachersGroup = doc.data()
+    const teachersGroup = doc.data();
     const students = await getStudentsOf(teachersGroup);
     const completeGroup = Object.assign(teachersGroup, {students});
-    teachersGroups.push(completeGroup)
+    teachersGroups.push(completeGroup);
   });
   return teachersGroups;
 }
@@ -166,92 +168,17 @@ export async function getAbsence(student) {
   });
 }
 
-export async function getAllAbsentees() {
-  const groupsIDs = await getGroupsIDs();
-  const groupsStudents = await getAllStudents(groupsIDs);
-  const filteredGroupsStudents = groupsStudents.filter(groupStudent => groupStudent.length).flat();
-  const attendants = await getAbsentStudents(filteredGroupsStudents);
-  const absentees = attendants.filter((attendant) => attendant.absences.length);
-  const absences = getAbsences(absentees);
-  return removeDuplicate(absences);
-}
-
 export function getStudentDocName(student) {
   const studentDocName = `${student.name.first}${student.name.last}${student.education}`;
   const filteredStudentDocName = studentDocName.replace(/\s+/g, '');
   return filteredStudentDocName;
 }
 
-async function getGroupsIDs() {
-  let groupsIDs = [];
-  await db.collection("groups").get().then((querySnapshot) => {
-    querySnapshot.forEach((doc) => {
-      groupsIDs.push(doc.id);
-    })
-  })
-  return groupsIDs;
-}
-
-async function getAllStudents(groupsIDs) {
-  const studentsGroups = await groupsIDs.map(async (groupID) => {
-    return await getStudentsOf({groupName: groupID});
-  });
-  return Promise.all(studentsGroups);
-}
-
-async function getAbsentStudents(groupStudents) {
-  const absentStudentLessons  = await getStudentLessons(groupStudents, "presence", "afwezig");
-  const unkownReasonAbsentLessons  = await getStudentLessons(groupStudents, "absence.reason", "overige");
-  const sickStudentLessons = await getStudentLessons(groupStudents, "absence.reason", "ziekte");
-  return absentStudentLessons.concat(unkownReasonAbsentLessons).concat(sickStudentLessons);
-}
-
-async function getStudentLessons(groupsStudents, field, value) {
-  const attendants = await groupsStudents.map(async (groupsStudent) => {
-    const student = Object.assign({}, groupsStudent);
-    const studentDocName = getStudentDocName(groupsStudent);
-    let absences = [];
-    await db.collection("groups").doc(groupsStudent.group).collection("students").doc(studentDocName).collection("lessons")
-      .where(field, "==", value)
-      .get()
-      .then((querySnapshot) => {
-        querySnapshot.forEach((doc) => {
-          const absence = doc.data();
-          absence.date = doc.id;
-          absences.push(absence);
-        })
-      })
-    student.absences = absences;
-    return student;
-  })
-  return Promise.all(attendants);
-}
-
-function getAbsences(absentees) {
-  return absentees.flatMap((absentee) => {
-    return absentee.absences.map((absence) => {
-      const student = Object.assign({}, absentee);
-      student.absence = absence;
-      student.date = absence.date;
-      delete student.absence.date;
-      return student;
-    })
-  });
-}
-
-function removeDuplicate(absences) {
-  return absences.filter((absence,index) => {
-    return index === absences.findIndex(obj => {
-      return JSON.stringify(obj) === JSON.stringify(absence);
-    });
-  });
-}
-
-async function getStudentsOf(teachersGroup) {
+export async function getStudentsOf(teachersGroup) {
   const querySnapshot = await db.collection("groups").doc(teachersGroup.groupName).collection('students').get();
   let students = [];
   querySnapshot.forEach((student) => {
-    students.push(student.data())
+    students.push(student.data());
   });
   return students;
 }
@@ -260,7 +187,7 @@ async function getGroup(groupName) {
   const querySnapshot = await db.collection("groups").where("groupName", "==", groupName).get();
   let groups = [];
   querySnapshot.forEach((doc) => {
-    groups.push(doc.id)
+    groups.push(doc.id);
   });
   return groups[0];
 }
